model-v1/requirements.txt
# Deepfake Detection Analysis - Dependencies
# Install with: pip install -r requirements.txt

# Deep Learning
torch>=2.0.0
transformers>=4.30.0

# Image Processing
Pillow>=9.0.0

# Data Processing
numpy>=1.24.0

# Machine Learning Metrics
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bar
tqdm>=4.65.0
